edubid-backend/requirements.txt
asgiref==3.9.2
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
Django==5.2.6
django-allauth==65.11.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.41.1
gunicorn==23.0.0
idna==3.10
pymysql==1.1.1
packaging==25.0
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
rsa==4.9.1
sendgrid==6.12.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0